split more time sensitive tests from race detection
